Added multi pick for both platforms
Dispose of popover Fixes #327

WIP - Basic implementation for multi pick on Android

Added multi-picker support for iOS

Added the customisations for ios picker UI and implemented UWP as single picker for now.

Added a customisation for the overlay image.

Added the ability to customise the titles on the ios picker.

Fix the incorrect orientation on iOS from multipicker

Fix for the orientation being incorrect when selecting a portrait photo from the galley

Updated the ELCImagePicker with the new version for Unified iOS

Make sure meta isn't null when passed into WriteImageToSavedPhotosAlbumAsync